requirements.txt
Pillow==2.6.1
Scrapy==0.24.4
Twisted==14.0.2
boto==2.34.0
cffi==0.8.6
characteristic==14.1.0
cryptography==0.6.1
cssselect==0.9.1
lxml==3.4.0
pyOpenSSL==0.14
pyasn1==0.1.7
pyasn1-modules==0.0.5
pycparser==2.10
queuelib==1.2.2
service-identity==14.0.0
six==1.8.0
w3lib==1.10.0
wsgiref==0.1.2
zope.interface==4.1.1
